requirements.txt
streamlit == 1.8.1
numpy == 1.22.3
tensorflow-cpu == 2.8
pillow == 9.0.1